Cursor for Turret Rotation, Tower Class and Firing Functionality
We can now rotate the direction of the turret using the mouse, Tower class was created for its own rotation and we have a bare bones Fire function that will be expanded upon. Tank rotation is still a bit iffy, the hit result when hovering over the tank makes it look and feel bad so need to fix that in future.